# frontend/openapi-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HamzaTex API",
    "description": "Textile ERP — staff mobile + admin backend. All endpoints require a JWT Bearer token except Auth.",
    "version": "v1"
  },
  "paths": {
...
    "/api/Device/register": {
      "post": {
        "tags": ["Device"],
        "summary": "Register or refresh an FCM device token for push notifications.",
        "operationId": "device_register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceViewModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
    "/api/Device/unregister": {
      "delete": {
        "tags": ["Device"],
        "summary": "Unregister a specific FCM device token (e.g. on logout from this device).",
        "operationId": "device_unregister",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterDeviceViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterDeviceViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterDeviceViewModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Expense/{id}": {
      "get": {
        "tags": ["Expense"],
        "summary": "Get an expense by ID.",
        "operationId": "expense_getById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseDtoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseDtoResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Expense"],
        "summary": "Update amount, mode, date, and notes. TransCategoryId cannot be changed — delete and re-create to reclassify.",
        "operationId": "expense_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseUpdateViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseUpdateViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseUpdateViewModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseDtoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseDtoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseDtoResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Expense"],
        "summary": "Hard delete an expense and its linked Transaction. Admin only.",
        "operationId": "expense_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Expense/pdf": {
      "get": {
        "tags": ["Expense"],
        "summary": "Export expenses as PDF. Admin sees all; non-admins see only their own. Optionally filter by type, mode, and date range.",
        "operationId": "expense_getPdf",
        "parameters": [
          {
            "name": "expenseTypeId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "modeId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Expense/{id}/pdf": {
      "get": {
        "tags": ["Expense"],
        "summary": "Download a single expense as a branded PDF receipt — type, amount, and details.",
        "operationId": "expense_getExpenseDossierPdf",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
    "/api/ExpenseType": {
      "post": {
        "tags": ["ExpenseType"],
        "summary": "Create a new expense type.",
        "operationId": "expenseType_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseTypeCreateViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseTypeCreateViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseTypeCreateViewModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseTypeDtoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseTypeDtoResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["ExpenseType"],
        "summary": "Get all expense types.",
        "operationId": "expenseType_getAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseTypeDtoListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ExpenseType/{id}": {
      "get": {
        "tags": ["ExpenseType"],
        "summary": "Get an expense type by ID.",
        "operationId": "expenseType_getById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseTypeDtoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseTypeDtoResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["ExpenseType"],
        "summary": "Update an expense type name.",
        "operationId": "expenseType_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseTypeUpdateViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseTypeUpdateViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseTypeUpdateViewModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseTypeDtoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseTypeDtoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseTypeDtoResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["ExpenseType"],
        "summary": "Delete an expense type. Fails if the type is seeded (ID 1 or 2) or has existing expenses.",
        "operationId": "expenseType_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExpenseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "expenseTypeId": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "example": 1
          },
          "expenseTypeName": {
            "type": "string",
            "nullable": true,
            "example": "string"
          },
          "amount": {
            "type": "number",
            "format": "double",
            "example": 150000
          },
          "transModeId": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "example": 1
          },
          "transModeName": {
            "type": "string",
            "nullable": true,
            "example": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "example": 1
          },
          "recordedByName": {
            "type": "string",
            "nullable": true,
            "example": "string"
          },
          "transCategoryId": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "example": 1
          },
          "transCategoryName": {
            "type": "string",
            "nullable": true,
            "example": "string"
          },
          "transactionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "example": 1
          },
          "expenseDate": {
            "type": "string",
            "format": "date",
            "example": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "example": "Optional notes here"
          },
          "createdAt": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ExpenseDtoResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "example": "Operation completed successfully"
          },
          "data": {
            "$ref": "#/components/schemas/ExpenseDto"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExpenseTypeCreateViewModel": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true,
            "example": "Ahmed Textiles"
          }
        },
        "additionalProperties": false
      },
      "ExpenseTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "name": {
            "type": "string",
            "nullable": true,
            "example": "Ahmed Textiles"
          },
          "createdAt": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "string"
          }
        },
        "additionalProperties": false
      },
      "ExpenseTypeDtoListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "example": "Operation completed successfully"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseTypeDto"
            },
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExpenseTypeDtoResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "example": "Operation completed successfully"
          },
          "data": {
            "$ref": "#/components/schemas/ExpenseTypeDto"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExpenseTypeUpdateViewModel": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true,
            "example": "Ahmed Textiles"
          }
        },
        "additionalProperties": false
      },
      "ExpenseUpdateViewModel": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double",
            "example": 150000
          },
          "transModeId": {
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "expenseDate": {
            "type": "string",
            "description": "Omit or set null to leave the existing date unchanged.",
            "format": "date",
            "nullable": true,
            "example": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "example": "Optional notes here"
          }
        },
        "additionalProperties": false
      },
...
      "RegisterDeviceViewModel": {
        "type": "object",
        "properties": {
          "pushToken": {
            "type": "string",
            "nullable": true,
            "example": "string"
          },
          "deviceType": {
            "type": "string",
            "nullable": true,
            "example": "string"
          },
          "appVersion": {
            "type": "string",
            "nullable": true,
            "example": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Response": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "example": "Operation completed successfully"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UnregisterDeviceViewModel": {
        "type": "object",
        "properties": {
          "pushToken": {
            "type": "string",
            "nullable": true,
            "example": "string"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter your JWT access token (without the 'Bearer ' prefix).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}